refactor(misc1): move line_breakcheck family of functions to os/input.c
