+++ src/nvim/autocmd.c
@@ -17,6 +17,7 @@
 #include "nvim/getchar.h"
 #include "nvim/misc1.h"
 #include "nvim/option.h"
+#include "nvim/os/input.h"
 #include "nvim/regexp.h"
 #include "nvim/search.h"
 #include "nvim/state.h"
