+++ src/nvim/regexp.c
@@ -65,6 +65,7 @@
 #include "nvim/memory.h"
 #include "nvim/message.h"
 #include "nvim/misc1.h"
+#include "nvim/os/input.h"
 #include "nvim/plines.h"
 #include "nvim/garray.h"
 #include "nvim/strings.h"
